Add create segment
