+++ src/Hilke.KineticConvolution/Tracing.cs
@@ -37,77 +37,5 @@ namespace Hilke.KineticConvolution
 
         public bool IsG1ContinuousWith(Tracing<TAlgebraicNumber> next) =>
             End == next.Start && EndDirection == next.StartDirection;
-
-        public static Tracing<TAlgebraicNumber> CreateArc(
-            IAlgebraicNumberCalculator<TAlgebraicNumber> calculator,
-            Fraction weight,
-            Point<TAlgebraicNumber> center,
-            DirectionRange<TAlgebraicNumber> directions,
-            TAlgebraicNumber radius)
-        {
-            if (center == null)
-            {
-                throw new ArgumentNullException(nameof(center));
-            }
-
-            if (directions == null)
-            {
-                throw new ArgumentNullException(nameof(directions));
-            }
-
-            if (radius == null)
-            {
-                throw new ArgumentNullException(nameof(radius));
-            }
-
-            var start = center.Translate(directions.Start, radius);
-            var end = center.Translate(directions.End, radius);
-
-            var startNormalDirection = directions.Start.NormalDirection();
-            var startDirection = directions.Orientation == Orientation.Clockwise
-                                     ? startNormalDirection.Opposite()
-                                     : startNormalDirection;
-
-            var endNormalDirection = directions.End.NormalDirection();
-            var endDirection = directions.Orientation == Orientation.Clockwise
-                                   ? endNormalDirection.Opposite()
-                                   : endNormalDirection;
-
-            return new Arc<TAlgebraicNumber>(
-                calculator,
-                weight,
-                center,
-                directions,
-                radius,
-                start,
-                end,
-                startDirection,
-                endDirection);
-        }
-
-        public static Tracing<TAlgebraicNumber> CreateSegment(
-            IAlgebraicNumberCalculator<TAlgebraicNumber> calculator,
-            Point<TAlgebraicNumber> start,
-            Point<TAlgebraicNumber> end,
-            Fraction weight)
-        {
-            if (start == null)
-            {
-                throw new ArgumentNullException(nameof(start));
-            }
-
-            if (end == null)
-            {
-                throw new ArgumentNullException(nameof(end));
-            }
-
-            if (start == end)
-            {
-                throw new ArgumentException("The start point cannot be the same as end point.", nameof(start));
-            }
-
-            var direction = start.DirectionTo(end);
-            return new Segment<TAlgebraicNumber>(calculator, start, end, direction, direction, weight);
-        }
     }
 }
